+++ config/environment.php
@@ -1,8 +1,4 @@
 <?php
-use Opulence\Environments\Environment;
-
-$environment = new Environment();
-
 /**
  * ----------------------------------------------------------
  * Load environment config files
@@ -11,12 +7,10 @@ $environment = new Environment();
  * Note:  For performance in production, it's highly suggested
  * you set environment variables on the server itself
  */
-foreach (glob(__DIR__ . "/environment/.env.*.php") as $environmentFile) {
-    if (basename($environmentFile) != ".env.example.php") {
-        require $environmentFile;
-    }
+$environmentConfigFiles = [
+    __DIR__ . "/environment/.env.app.php"
+];
+
+foreach ($environmentConfigFiles as $environmentFile) {
+    require $environmentFile;
 }
-
-$environment->setName($environment->getVar("ENV_NAME") ?? Environment::PRODUCTION);
-
-return $environment;