+++ config/http/sessions.php
@@ -1,5 +1,6 @@
 <?php
 use Opulence\Cache\FileBridge;
+use Opulence\Environments\Environment;
 use Opulence\Framework\Configuration\Config;
 use Opulence\Sessions\Handlers\FileSessionHandler;
 
@@ -19,7 +20,7 @@ return [
      * "name" => The name of the session
      * "isEncrypted" => Whether or not the session data is encrypted when stored
      */
-    "handler" => $environment->getVar("SESSION_HANDLER", FileSessionHandler::class),
+    "handler" => Environment::getVar("SESSION_HANDLER", FileSessionHandler::class),
     "lifetime" => 7200,
     "name" => "__opulence_session",
     "isEncrypted" => false,
@@ -47,10 +48,10 @@ return [
      * "cookie.isSecure" => Whether or not the cookie is secure
      * "cookie.path" => The path of the cookie
      */
-    "cookie.domain" => $environment->getVar("SESSION_COOKIE_DOMAIN", ""),
+    "cookie.domain" => Environment::getVar("SESSION_COOKIE_DOMAIN", ""),
     "cookie.isHttpOnly" => true,
-    "cookie.isSecure" => $environment->getVar("SESSION_COOKIE_IS_SECURE", false),
-    "cookie.path" => $environment->getVar("SESSION_COOKIE_PATH", "/"),
+    "cookie.isSecure" => Environment::getVar("SESSION_COOKIE_IS_SECURE", false),
+    "cookie.path" => Environment::getVar("SESSION_COOKIE_PATH", "/"),
 
     /**
      * ----------------------------------------------------------
@@ -61,7 +62,7 @@ return [
      * "cache.clientName" => The name of the client to use in your cache bridge
      * "cache.keyPrefix" => The prefix to use on all cache keys to avoid naming collisions
      */
-    "cache.bridge" => $environment->getVar("SESSION_CACHE_BRIDGE", FileBridge::class),
+    "cache.bridge" => Environment::getVar("SESSION_CACHE_BRIDGE", FileBridge::class),
     "cache.clientName" => "default",
     "cache.keyPrefix" => "opulence:",