+++ tests/src/Project/Application/Console/IntegrationTestCase.php
@@ -22,8 +22,8 @@ class IntegrationTestCase extends BaseIntegrationTestCase
      */
     public function setUp()
     {
-        $paths = require __DIR__ . "/../../../../../config/paths.php";
-        $this->environment = require __DIR__ . "/../../../../../config/environment.php";
+        require __DIR__ . "/../../../../../config/paths.php";
+        require __DIR__ . "/../../../../../config/environment.php";
         $this->application = require __DIR__ . "/../../../../../config/application.php";
         /** @var IContainer $container */
         $this->container = $container;